refactor: split up ext
